backend/requirements.txt
# Core data
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0          # Excel support
xlrd>=2.0.1              # Legacy .xls support

# Visualization (chart data for frontend)
plotly>=5.18.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9  # File uploads

# AI / LLM
openai>=1.12.0

# Supabase
supabase>=2.3.0

# Utilities
python-dotenv>=1.0.0
scipy>=1.11.0            # Statistical checks
statsmodels>=0.14.0      # OLS trendlines in scatter plots
